fix: storage Вынести FileStorage в отдельный файл
